Added base year car ownership model to the Silo model container. Implemented a driver license ownership model to create license for new persons and update driver license status of existing persons. Base year car ownership model is now applied to create new households that come up as a result of people inmigration, marriage, divorce as well as people leaving their parental households.

+++ src/main/java/de/tum/bgu/msm/syntheticPopulationGenerator/CreateCarOwnershipJSCalculator.java
@@ -25,7 +25,7 @@ public class CreateCarOwnershipJSCalculator extends JavaScriptCalculator<double[
         this.bindings.put("income", income);
     }
 
-    public void setLogDistanceToTransit (int logDistanceToTransit) {
+    public void setLogDistanceToTransit (double logDistanceToTransit) {
         this.bindings.put("logDistanceToTransit", logDistanceToTransit);
     }
